@@ -53,7 +53,7 @@
     "require": {
         "paragonie/constant_time_encoding": "^1|^2",
         "paragonie/random_compat": "^1.4|^2.0",
-        "php": ">=5.3.3"
+        "php": ">=7.0"
     },
     "require-dev": {
         "phing/phing": "~2.7",
