@@ -24,14 +24,55 @@
 
 namespace phpseclib\Crypt\RSA;
 
-use ParagonIE\ConstantTime\Base64;
-use ParagonIE\ConstantTime\Hex;
-use phpseclib\Crypt\AES;
-use phpseclib\Crypt\DES;
-use phpseclib\Crypt\Random;
-use phpseclib\Crypt\TripleDES;
 use phpseclib\Math\BigInteger;
-use phpseclib\Common\Functions\ASN1;
+use phpseclib\Crypt\Common\PKCS1 as Progenitor;
+use phpseclib\File\ASN1;
+
+// version must be multi if otherPrimeInfos present
+define(__NAMESPACE__ . '\Version', [
+    'type'    => ASN1::TYPE_INTEGER,
+    'mapping' => ['two-prime', 'multi']
+]);
+
+define(__NAMESPACE__ . '\OtherPrimeInfo', [
+    'type' => ASN1::TYPE_SEQUENCE,
+    'children' => [
+        'prime' =>       ['type' => ASN1::TYPE_INTEGER], // ri
+        'exponent' =>    ['type' => ASN1::TYPE_INTEGER], // di
+        'coefficient' => ['type' => ASN1::TYPE_INTEGER]  // ti
+    ]
+]);
+
+define(__NAMESPACE__ . '\OtherPrimeInfos', [
+    'type' => ASN1::TYPE_SEQUENCE,
+    'min' => 1,
+    'max' => -1,
+    'children' => OtherPrimeInfo
+]);
+
+define(__NAMESPACE__ . '\RSAPrivateKey', [
+    'type' => ASN1::TYPE_SEQUENCE,
+    'children' => [
+        'version' =>         Version,
+        'modulus' =>         ['type' => ASN1::TYPE_INTEGER], // n
+        'publicExponent' =>  ['type' => ASN1::TYPE_INTEGER], // e
+        'privateExponent' => ['type' => ASN1::TYPE_INTEGER], // d
+        'prime1' =>          ['type' => ASN1::TYPE_INTEGER], // p
+        'prime2' =>          ['type' => ASN1::TYPE_INTEGER], // q
+        'exponent1' =>       ['type' => ASN1::TYPE_INTEGER], // d mod (p-1)
+        'exponent2' =>       ['type' => ASN1::TYPE_INTEGER], // d mod (q-1)
+        'coefficient' =>     ['type' => ASN1::TYPE_INTEGER], // (inverse of q) mod p
+        'otherPrimeInfos' => OtherPrimeInfos + ['optional' => true]
+    ]
+]);
+
+define(__NAMESPACE__ . '\RSAPublicKey', [
+    'type' => ASN1::TYPE_SEQUENCE,
+    'children' => [
+        'modulus' =>         ['type' => ASN1::TYPE_INTEGER],
+        'publicExponent' =>  ['type' => ASN1::TYPE_INTEGER]
+    ]
+]);
 
 /**
  * PKCS#1 Formatted RSA Key Handler
@@ -40,25 +81,58 @@ use phpseclib\Common\Functions\ASN1;
  * @author  Jim Wigginton <terrafrost@php.net>
  * @access  public
  */
-class PKCS1 extends PKCS
+class PKCS1 extends Progenitor
 {
     /**
-     * Default encryption algorithm
-     *
-     * @var string
-     * @access private
-     */
-    static $defaultEncryptionAlgorithm = 'DES-EDE3-CBC';
-
-    /**
-     * Sets the default encryption algorithm
+     * Break a public or private key down into its constituent components
      *
      * @access public
-     * @param string $algo
+     * @param string $key
+     * @param string $password optional
+     * @return array
      */
-    static function setEncryptionAlgorithm($algo)
+    static function load($key, $password = '')
     {
-        self::$defaultEncryptionAlgorithm = $algo;
+        if (!is_string($key)) {
+            return false;
+        }
+
+        $components = ['isPublicKey' => strpos($key, 'PUBLIC') !== false];
+
+        $key = parent::load($key, $password);
+        if ($key === false) {
+            return false;
+        }
+
+        $asn1 = new ASN1();
+        $decoded = $asn1->decodeBER($key);
+        if (empty($decoded)) {
+            return false;
+        }
+
+        $key = $asn1->asn1map($decoded[0], RSAPrivateKey);
+        if (is_array($key)) {
+            $components+= [
+                'modulus' => $key['modulus'],
+                'publicExponent' => $key['publicExponent'],
+                'privateExponent' => $key['privateExponent'],
+                'primes' => [1 => $key['prime1'], $key['prime2']],
+                'exponents' => [1 => $key['exponent1'], $key['exponent2']],
+                'coefficients' => [2 => $key['coefficient']]
+            ];
+            if ($key['version'] == 'multi') {
+                foreach ($key['otherPrimeInfos'] as $primeInfo) {
+                    $components['primes'][] = $primeInfo['prime'];
+                    $components['exponents'][] = $primeInfo['exponent'];
+                    $components['coefficients'][] = $primeInfo['coefficient'];
+                }
+            }
+            return $components;
+        }
+
+        $key = $asn1->asn1map($decoded[0], RSAPublicKey);
+
+        return is_array($key) ? $components + $key : false;
     }
 
     /**
@@ -77,64 +151,29 @@ class PKCS1 extends PKCS
     static function savePrivateKey(BigInteger $n, BigInteger $e, BigInteger $d, $primes, $exponents, $coefficients, $password = '')
     {
         $num_primes = count($primes);
-        $raw = array(
-            'version' => $num_primes == 2 ? chr(0) : chr(1), // two-prime vs. multi
-            'modulus' => $n->toBytes(true),
-            'publicExponent' => $e->toBytes(true),
-            'privateExponent' => $d->toBytes(true),
-            'prime1' => $primes[1]->toBytes(true),
-            'prime2' => $primes[2]->toBytes(true),
-            'exponent1' => $exponents[1]->toBytes(true),
-            'exponent2' => $exponents[2]->toBytes(true),
-            'coefficient' => $coefficients[2]->toBytes(true)
-        );
-
-        $components = array();
-        foreach ($raw as $name => $value) {
-            $components[$name] = pack('Ca*a*', self::ASN1_INTEGER, ASN1::encodeLength(strlen($value)), $value);
+        $key = [
+            'version' => $num_primes == 2 ? 'two-prime' : 'multi',
+            'modulus' => $n,
+            'publicExponent' => $e,
+            'privateExponent' => $d,
+            'prime1' => $primes[1],
+            'prime2' => $primes[2],
+            'exponent1' => $exponents[1],
+            'exponent2' => $exponents[2],
+            'coefficient' => $coefficients[2]
+        ];
+        for ($i = 3; $i <= $num_primes; $i++) {
+            $key['otherPrimeInfos'][] = [
+                'prime' => $primes[$i],
+                'exponent' => $exponents[$i],
+                'coefficient' => $coefficients[$i]
+            ];
         }
 
-        $RSAPrivateKey = implode('', $components);
+        $asn1 = new ASN1();
+        $key = $asn1->encodeDER($key, RSAPrivateKey);
 
-        if ($num_primes > 2) {
-            $OtherPrimeInfos = '';
-            for ($i = 3; $i <= $num_primes; $i++) {
-                // OtherPrimeInfos ::= SEQUENCE SIZE(1..MAX) OF OtherPrimeInfo
-                //
-                // OtherPrimeInfo ::= SEQUENCE {
-                //     prime             INTEGER,  -- ri
-                //     exponent          INTEGER,  -- di
-                //     coefficient       INTEGER   -- ti
-                // }
-                $OtherPrimeInfo = pack('Ca*a*', self::ASN1_INTEGER, ASN1::encodeLength(strlen($primes[$i]->toBytes(true))), $primes[$i]->toBytes(true));
-                $OtherPrimeInfo.= pack('Ca*a*', self::ASN1_INTEGER, ASN1::encodeLength(strlen($exponents[$i]->toBytes(true))), $exponents[$i]->toBytes(true));
-                $OtherPrimeInfo.= pack('Ca*a*', self::ASN1_INTEGER, ASN1::encodeLength(strlen($coefficients[$i]->toBytes(true))), $coefficients[$i]->toBytes(true));
-                $OtherPrimeInfos.= pack('Ca*a*', self::ASN1_SEQUENCE, ASN1::encodeLength(strlen($OtherPrimeInfo)), $OtherPrimeInfo);
-            }
-            $RSAPrivateKey.= pack('Ca*a*', self::ASN1_SEQUENCE, ASN1::encodeLength(strlen($OtherPrimeInfos)), $OtherPrimeInfos);
-        }
-
-        $RSAPrivateKey = pack('Ca*a*', self::ASN1_SEQUENCE, ASN1::encodeLength(strlen($RSAPrivateKey)), $RSAPrivateKey);
-
-        if (!empty($password) || is_string($password)) {
-            $cipher = self::getEncryptionObject(self::$defaultEncryptionAlgorithm);
-            $iv = Random::string($cipher->getBlockLength() >> 3);
-            $cipher->setKey(self::generateSymmetricKey($password, $iv, $cipher->getKeyLength() >> 3));
-            $cipher->setIV($iv);
-            $iv = strtoupper(Hex::encode($iv));
-            $RSAPrivateKey = "-----BEGIN RSA PRIVATE KEY-----\r\n" .
-                     "Proc-Type: 4,ENCRYPTED\r\n" .
-                     "DEK-Info: " . self::$defaultEncryptionAlgorithm . ",$iv\r\n" .
-                     "\r\n" .
-                     chunk_split(Base64::encode($cipher->encrypt($RSAPrivateKey)), 64) .
-                     '-----END RSA PRIVATE KEY-----';
-        } else {
-            $RSAPrivateKey = "-----BEGIN RSA PRIVATE KEY-----\r\n" .
-                     chunk_split(Base64::encode($RSAPrivateKey), 64) .
-                     '-----END RSA PRIVATE KEY-----';
-        }
-
-        return $RSAPrivateKey;
+        return self::wrapPrivateKey($key, 'RSA', $password);
     }
 
     /**
@@ -147,31 +186,14 @@ class PKCS1 extends PKCS
      */
     static function savePublicKey(BigInteger $n, BigInteger $e)
     {
-        $modulus = $n->toBytes(true);
-        $publicExponent = $e->toBytes(true);
+        $key = [
+            'modulus' => $n,
+            'publicExponent' => $e
+        ];
 
-        // from <http://tools.ietf.org/html/rfc3447#appendix-A.1.1>:
-        // RSAPublicKey ::= SEQUENCE {
-        //     modulus           INTEGER,  -- n
-        //     publicExponent    INTEGER   -- e
-        // }
-        $components = array(
-            'modulus' => pack('Ca*a*', self::ASN1_INTEGER, ASN1::encodeLength(strlen($modulus)), $modulus),
-            'publicExponent' => pack('Ca*a*', self::ASN1_INTEGER, ASN1::encodeLength(strlen($publicExponent)), $publicExponent)
-        );
+        $asn1 = new ASN1();
+        $key = $asn1->encodeDER($key, RSAPublicKey);
 
-        $RSAPublicKey = pack(
-            'Ca*a*a*',
-            self::ASN1_SEQUENCE,
-            ASN1::encodeLength(strlen($components['modulus']) + strlen($components['publicExponent'])),
-            $components['modulus'],
-            $components['publicExponent']
-        );
-
-        $RSAPublicKey = "-----BEGIN RSA PUBLIC KEY-----\r\n" .
-                        chunk_split(Base64::encode($RSAPublicKey), 64) .
-                        '-----END RSA PUBLIC KEY-----';
-
-        return $RSAPublicKey;
+        return self::wrapPublicKey($key, 'RSA');
     }
 }
