@@ -1511,7 +1511,7 @@ class SSH2
 
             -- http://tools.ietf.org/html/rfc4419#section-6.2 */
             $one = new BigInteger(1);
-            $keyLength = min($kexHash->getLength(), max($encryptKeyLength, $decryptKeyLength));
+            $keyLength = min($kexHash->getLengthInBytes(), max($encryptKeyLength, $decryptKeyLength));
             $max = $one->bitwise_leftShift(16 * $keyLength); // 2 * 8 * $keyLength
             $max = $max->subtract($one);
 
@@ -2927,9 +2927,6 @@ class SSH2
 
         if ($this->decrypt !== false) {
             $raw = $this->decrypt->decrypt($raw);
-        }
-        if ($raw === false) {
-            throw new \RuntimeException('Unable to decrypt content');
         }
 
         extract(unpack('Npacket_length/Cpadding_length', Strings::shift($raw, 5)));
